feat!(node-stdlib): Rename Box to Ref
Move Ref to global.ts

BREAKING CHANGE: Box is now Ref
