+++ packages/node-stdlib/src/core/mod.ts
@@ -1,5 +1,4 @@
 // Copyright (c) 2020 Christopher Szatmary <cs@christopherszatmary.com>
 // All rights reserved. MIT License.
 
-export * from "./box";
 export * from "./result";
