+++ packages/node-stdlib/src/global.ts
@@ -10,7 +10,7 @@ This file should not import any other files from this library since it
 should be able to be imported by any other file.
 */
 
-import { inspect } from "util";
+import { inspect, InspectOptions } from "util";
 
 declare global {
   /**
@@ -110,5 +110,49 @@ export function recover(e: unknown): unknown {
 export function* range(lower: number, upper: number, step = 1): Generator<number, void> {
   for (let i = lower; i < upper; i += step) {
     yield i;
+  }
+}
+
+/**
+ * `Ref` provides a way to wrap a value so that it can be used as a reference.
+ * This allows functionality similar to that of references or pointers from
+ * other languages.
+ */
+export class Ref<T> {
+  #value: T;
+
+  constructor(value: T) {
+    this.#value = value;
+  }
+
+  /**
+   * Sets the value in the Ref to the given `value`.
+   */
+  set(value: T): void {
+    this.#value = value;
+  }
+
+  /**
+   * Returns the value stored inside the Ref.
+   */
+  deref(): T {
+    return this.#value;
+  }
+
+  /**
+   * Custom inspect implementation for use with node's `util.inspect`.
+   */
+  [inspect.custom](depth?: number | null, options?: InspectOptions): string {
+    if (depth == null || depth < 0) {
+      return "Ref {}";
+    }
+
+    const newOpts = {
+      ...options,
+      depth: options?.depth == null ? null : options.depth - 1,
+    };
+
+    const value = inspect(this.#value, newOpts);
+    return `Ref { ${value} }`;
   }
 }
